pyproject.toml
[project]
name = "jailbreakforge"
version = "0.1.0"
description = "AI Security Red Teaming Framework — Automated jailbreak testing for AI models"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Nitram2704"}
]

dependencies = [
    "typer>=0.15.1",
    "python-dotenv>=1.0",
    "rich>=13.0",
    "httpx>=0.27",
    "pyyaml>=6.0",
    "jinja2>=3.1",
]

[project.scripts]
jailbreakforge = "jailbreakforge.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/jailbreakforge"]

[tool.pytest.ini_options]
testpaths = ["tests"]
